requirements.txt
folium>=0.14.0
pandas>=1.5.0
requests>=2.28.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
jupyter>=1.0.0
notebook>=6.4.0
geopandas>=0.12.0
gspread>=5.0.0
google-auth>=2.0.0
